src/regexp.ts
export interface Flags {
	dotAll: boolean;
	global: boolean;
	ignoreCase: boolean;
	multiline: boolean;
	unicode: boolean;
}

export class RegExp {
	public static parseFlagsArray(flags: string): Flags {
		// TODO This should check that the string is in the format /[sgimu]*/
		return {
			dotAll: flags.includes("s"),
			global: flags.includes("g"),
			ignoreCase: flags.includes("i"),
			multiline: flags.includes("m"),
			unicode: flags.includes("u"),
		};
	}

	public readonly flags: Flags;
	public readonly pattern: string;

	public constructor(private readonly regExp: string) {
		// TODO This should check that the string is in the format /\/.*\/[a-z]*/
		const flagIndex = regExp.lastIndexOf("/");
		const flagString = regExp.substr(flagIndex);
		this.flags = RegExp.parseFlagsArray(flagString);
		this.pattern = regExp.substr(1, flagIndex - 1);
	}
}
